add delete all data functionality in PrivacyScreen instead
